Fix segfault issue on Arch
Somehow the LCL causes a segfault when it tries to get the monitor's DPI
on Arch. This seems to be resolved by letting Lazarus update the forms
(by slightly moving them around in the IDE, it reorganizes a bunch of
properties and updates the LCL version.... which seems to fix things). I
don't understand it, but oh well.

 https://gitlab.com/freepascal.org/lazarus/lazarus/-/issues/41025