Ship FFMpeg and RGBDS with the Linux build
This commit adds those binaries to the zip file we ship on GitHub. It
still doesn't come with the libs (should use AppImage for that maybe?)
so it's not fully self contained, but this should at least solve the
missing rgbds problem a few linux users have had.

Squashed commit of the following:

commit 1a43fc9f1020ef1c6509121e390b1f2b74b9a9ed
Author: Nick Faro <yux50000@hotmail.com>
Date:   Sun Jul 14 22:58:02 2024 +0400

    Fix packaging again

commit ff6902d6ea9b0ef406866681a111e9d15f5ea998
Author: Nick Faro <yux50000@hotmail.com>
Date:   Sun Jul 14 18:44:09 2024 +0400

    fix packaging step

commit e23d35101b41529027231ee3de446d27e495a0a6
Author: Nick Faro <yux50000@hotmail.com>
Date:   Sun Jul 14 18:36:46 2024 +0400

    Install necessary deps

commit c146630daf8f75a46fd1e5b58c7c09a226149940
Author: Nick Faro <yux50000@hotmail.com>
Date:   Sun Jul 14 18:29:14 2024 +0400

    WIP shipping ffmpeg and rgbds binaries with HT on linux